Refactor navigation health check and add new URLs
Refactor navigation health check and add new URLs

This commit refactors the navigation health check code and adds new URLs to be checked. The `get_urls` and `post_urls` arrays have been added to the `navigation` section of the `config.yml` file. The health check now makes GET requests to all the URLs in the `get_urls` array and POST requests to all the URLs in the `post_urls` array. Additionally, the `put_urls` array has been added to the `navigation` section, although it is currently empty.

The unused imports and function calls related to server and service health checks have been commented out in the `main.go` file.

The `health.go` file in the `navigation` package has been modified to handle the new URLs. The health check now makes GET requests to all the URLs in the `get_urls` array and POST requests to all the URLs in the `post_urls` array. The response from each request is checked for a status of "1" and appropriate alerts are sent.

Note: The commit message does not include any issue references, tags, or author names. These can be added when committing the changes.